Create AEAD sub-module, separate modules for algorithms.
I agree to license my contributions to each file under the
same terms given at the top of each file I changed.
